requirements.txt
google.genai
GitPython
python-dotenv
streamlit 
streamlit-mermaid